+++ src/main/java/me/opl/apps/modelcreator2/util/RotationHelper.java
@@ -12,11 +12,11 @@ public class RotationHelper {
 	 * @param origin Point to rotate around, ignored if <i>null</i>
 	 * @return Result of rotation as a new position object
 	 */
-	public static Position rotateX(Position point, double angle, Position origin) {
-		if (angle == 0d) return point.clone();
+	public static Position rotateX(Position point, float angle, Position origin) {
+		if (angle == 0f) return point.clone();
 
-		double sin = Math.sin(angle);
-		double cos = Math.cos(angle);
+		float sin = (float) Math.sin(angle);
+		float cos = (float) Math.cos(angle);
 
 		Position p = point.clone();
 		if (origin != null) p.subtract(origin);
@@ -25,11 +25,11 @@ public class RotationHelper {
 		return p;
 	}
 
-	public static Position rotateX(Position point, double angle) {
-		if (angle == 0d) return point.clone();
+	public static Position rotateX(Position point, float angle) {
+		if (angle == 0f) return point.clone();
 
-		double sin = Math.sin(angle);
-		double cos = Math.cos(angle);
+		float sin = (float) Math.sin(angle);
+		float cos = (float) Math.cos(angle);
 
 		Position p = point.clone();
 		p.set(p.getX(), p.getY() * cos + p.getZ() * sin, p.getY() * -sin + p.getZ() * cos);
@@ -44,22 +44,22 @@ public class RotationHelper {
 	 * @param origin Point to rotate around
 	 * @return Result of rotation as a new position object
 	 */
-	public static Position rotateY(Position point, double angle, Position origin) {
-		if (angle == 0d) return point.clone();
+	public static Position rotateY(Position point, float angle, Position origin) {
+		if (angle == 0f) return point.clone();
 
-		double sin = Math.sin(angle);
-		double cos = Math.cos(angle);
+		float sin = (float) Math.sin(angle);
+		float cos = (float) Math.cos(angle);
 
 		Position p = point.clone().subtract(origin);
 		p.set(p.getX() * cos + p.getZ() * -sin, p.getY(), p.getX() * sin + p.getZ() * cos);
 		return p.add(origin);
 	}
 
-	public static Position rotateY(Position point, double angle) {
-		if (angle == 0d) return point.clone();
+	public static Position rotateY(Position point, float angle) {
+		if (angle == 0f) return point.clone();
 
-		double sin = Math.sin(angle);
-		double cos = Math.cos(angle);
+		float sin = (float) Math.sin(angle);
+		float cos = (float) Math.cos(angle);
 
 		Position p = point.clone();
 		p.set(p.getX() * cos + p.getZ() * -sin, p.getY(), p.getX() * sin + p.getZ() * cos);
@@ -74,25 +74,25 @@ public class RotationHelper {
 	 * @param origin Point to rotate around
 	 * @return Result of rotation as a new position object
 	 */
-	public static Position rotateZ(Position point, double angle, Position origin) {
-		double sin = Math.sin(angle);
-		double cos = Math.cos(angle);
+	public static Position rotateZ(Position point, float angle, Position origin) {
+		float sin = (float) Math.sin(angle);
+		float cos = (float) Math.cos(angle);
 
 		Position p = point.clone().subtract(origin);
 		p.set(p.getX() * cos + p.getY() * sin, p.getX() * -sin + p.getY() * cos, p.getZ());
 		return p.add(origin);
 	}
 
-	public static Position rotateZ(Position point, double angle) {
-		double sin = Math.sin(angle);
-		double cos = Math.cos(angle);
+	public static Position rotateZ(Position point, float angle) {
+		float sin = (float) Math.sin(angle);
+		float cos = (float) Math.cos(angle);
 
 		Position p = point.clone();
 		p.set(p.getX() * cos + p.getY() * sin, p.getX() * -sin + p.getY() * cos, p.getZ());
 		return p;
 	}
 
-	public static Position rotate(Position point, Axis axis, double angle) {
+	public static Position rotate(Position point, Axis axis, float angle) {
 		Position p = point.clone();
 
 		if (axis == Axis.X) rotateX(p, angle);
@@ -102,7 +102,7 @@ public class RotationHelper {
 		return p;
 	}
 
-	public static Position rotate(Position point, Axis axis, double angle, Position origin) {
+	public static Position rotate(Position point, Axis axis, float angle, Position origin) {
 		Position p = point.clone();
 
 		if (axis == Axis.X) rotateX(p, angle, origin);
